pyproject.toml
[project]
name = "finaid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "llama-index>=0.12.32",
    "llama-index-embeddings-huggingface>=0.5.3",
    "llama-index-multi-modal-llms-gemini>=0.5.0",
    "llama-index-vector-stores-qdrant>=0.6.0",
    "pillow>=10.4.0",
    "qdrant-client>=1.14.1",
    "streamlit>=1.44.1",
]
